feat: add bounded semantic relation foundations
- Add persisted relation facts for bindings, bases, assignments, and macros
- Resolve typed member, callback, function-pointer, and macro call candidates conservatively
- Expose entity-bound references with read/write roles and bidirectional inheritance queries
- Integrate saved and unsaved relation facts with bounded scans, cancellation, and coverage diagnostics
- Add relation foundation tests, benchmarking, schema migration, and documentation

+++ crates/fossilsense/src/store.rs
@@ -85,7 +85,7 @@ fn read_language_evidence(
 
 #[derive(Debug, Clone, Copy)]
 pub struct PersistenceDiagnostics {
-    pub fact_mask: u8,
+    pub fact_mask: u16,
     pub parse_error_count: usize,
     pub fallback_used: bool,
 }
